refactor reporting snapshot routes
